+++ src/registry/domain/components-cache/index.js
@@ -1,58 +1,16 @@
 'use strict';
 
-const async = require('async');
-const semver = require('semver');
 const _ = require('underscore');
 
-const eventsHandler = require('./events-handler');
-const getUnixUTCTimestamp = require('../../utils/get-unix-utc-timestamp');
+const ComponentsList = require('./components-list');
+const eventsHandler = require('../events-handler');
+const getUnixUTCTimestamp = require('../../../utils/get-unix-utc-timestamp');
 
 module.exports = function(conf, cdn){
 
   let cachedComponentsList, refreshLoop;
 
-  const componentsList = {
-    getFromJson: callback => cdn.getJson(`${conf.s3.componentsDir}/components.json`, true, callback),
-
-    getFromDirectories: (callback) => {
-      const componentsInfo = {};
-
-      const getVersionsForComponent = (componentName, cb) => {
-        cdn.listSubDirectories(`${conf.s3.componentsDir}/${componentName}`, (err, versions) => {
-          if(err){ return cb(err); }
-          cb(null, versions.sort(semver.compare));
-        });
-      };
-
-      cdn.listSubDirectories(conf.s3.componentsDir, (err, components) => {
-        if(err){
-          if(err.code === 'dir_not_found'){
-            return callback(null, {
-              lastEdit: getUnixUTCTimestamp(),
-              components: []
-            });
-          }
-
-          return callback(err);
-        }
-
-        async.map(components, getVersionsForComponent, (errors, versions) => {
-          if(errors){ return callback(errors); }
-
-          _.forEach(components, (component, i) => {
-            componentsInfo[component] = versions[i];
-          });
-
-          callback(null, {
-            lastEdit: getUnixUTCTimestamp(),
-            components: componentsInfo
-          });
-        });
-      });
-    },
-
-    save: (data, callback) => cdn.putFileContent(JSON.stringify(data), `${conf.s3.componentsDir}/components.json`, true, callback)
-  };
+  const componentsList = ComponentsList(conf, cdn);
 
   const poll = () => setTimeout(() => {
     componentsList.getFromJson((err, data) => {
